Migrate main API to app route handler
