+++ src/lib/query/index.ts
@@ -1,6 +1,5 @@
 import parseQuery from "./parseQuery";
 import QueryParams from "./QueryParams";
-import rawQuery from "./rawQuery";
 
 export type { QueryParams };
-export { rawQuery, parseQuery };
+export { parseQuery };
